\name{busDaysOffset}
\alias{busDaysOffset}
\title{function to pass R objects to compiled C code loaded into R}
\usage{
busDaysOffset(fromDate, offset, holidays = NULL)
}
\arguments{
  \item{fromDate}{starting date}

  \item{offset}{function}

  \item{holidays}{holidays is an input for holiday files to
  adjust to business day counting. Default is \code{None}.}
}
\description{
function to pass R objects to compiled C code loaded into R
}

