% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.wt.bubbles.R
\name{calc.wt.bubbles}
\alias{calc.wt.bubbles}
\title{Calculate the Weights for Bubble Plots}
\usage{
calc.wt.bubbles(dat, grp, q, fact = 0.12)
}
\arguments{
\item{dat}{A data set from which to derive the relative frequencies}

\item{grp}{A vector giving the group memberships.}

\item{q}{An integer such that the rating scale is \code{1:q}.}

\item{fact}{A schrinkage factor.}
}
\description{
Calculate weights for the bubbles in the plot method of \code{cds} objects. The relative 
frequencies within a dset of groups are used to calculate the size of the bubble so that the area
of the bubble is proportional to the relative frequency of the rating category within that group.
}
\author{
Pieter Schoonees
}
\keyword{utility}
