% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_cds.R
\name{cds-package}
\alias{cds-package}
\title{Constrained Dual Scaling for Successive Categories}
\description{
Fit constrained dual scaling for detecting response styles.
}
\references{
Departmental report available

Schoonees, P.C., Velden, M. van de & Groenen, P.J.F. (2013).
Constrained Dual Scaling for Detecting Response Styles in Categorical Data.
(EI report series EI 2013-10). Rotterdam: Econometric Institute.
}
\author{
Pieter C. Schoonees
}
\keyword{package}
