% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_version.R
\name{CDS_version}
\alias{CDS_version}
\title{Get metadata on current CDS version}
\usage{
CDS_version()
}
\value{
Dataframe containing current CDS version number, build date, and code version.
}
\description{
Return metadata about the current CDS version
}
\examples{
{
CDS_version_metadata <- CDS_version()
}

}
