\name{CDSdf}
\alias{CDSdf}
\title{function to convert relevant slots in object of class CDS into a data frame}
\usage{
CDSdf(object)
}
\arguments{
  \item{object}{object of class CDS}
}
\value{
returns data frame with relevant data from the CDS function
}
\description{
function to convert relevant slots in object of class CDS
into a data frame
}

