% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdstoclue.R
\name{cl_class_ids.cds}
\alias{cl_class_ids.cds}
\alias{is.cl_partition.cds}
\alias{is.cl_hard_partition.cds}
\alias{cl_class_ids.cdsdata}
\alias{is.cl_partition.cdsdata}
\alias{is.cl_hard_partition.cdsdata}
\title{S3 Methods for Integration into \pkg{clue} Framework}
\usage{
\method{cl_class_ids}{cds}(x)

\method{is.cl_partition}{cds}(x)

\method{is.cl_hard_partition}{cds}(x)

\method{cl_class_ids}{cdsdata}(x)

\method{is.cl_partition}{cdsdata}(x)

\method{is.cl_hard_partition}{cdsdata}(x)
}
\arguments{
\item{x}{An object of class \code{cds}}
}
\description{
These methods integrate the class \code{cds} into the framwork set out in
package \pkg{clue}. Use can therefore by made of \code{\link[clue:cl_agreement]{cl_agreement}} to
calculate concordance measures between different solutions.
}
