% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.scales.R, R/clean.scales.cds.R,
%   R/clean.scales.cdslist.R
\name{clean.scales}
\alias{clean.scales}
\alias{clean.scales.cds}
\alias{clean.scales.cdslist}
\title{Impute Optimal Scores for Rating Categories}
\usage{
clean.scales(object, data, K, col.subset = NULL, ...)

\method{clean.scales}{cds}(object, data, K, col.subset = NULL, ...)

\method{clean.scales}{cdslist}(object, data, K, col.subset = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{cds}}

\item{data}{An object of class \code{cdsdata} to be cleaned, or the original data.}

\item{K}{The number of classes in the solution that must be kept.}

\item{col.subset}{An optional subset}

\item{\dots}{Additional arguments.}
}
\description{
Replace original ratings with optimal scores based on \code{\link{cds}} output..
}
\keyword{utility}
