\name{coercetoChar}
\alias{coercetoChar}
\title{check if argument is not a character and coerce it to character}
\usage{
coercetoChar(x)
}
\arguments{
  \item{x}{input into the function}
}
\value{
true if it is a character
}
\description{
check if argument is not a character and coerce it to
character
}

