% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS.R
\name{compQ0}
\alias{compQ0}
\title{xxx}
\usage{
compQ0(m1, s1, m2, s2, TAN_ALPHA_0)
}
\arguments{
\item{m1}{xxx}

\item{s1}{xxx}

\item{m2}{xxx}

\item{TAN_ALPHA_0}{xxx}

\item{m2}{xxx}
}
\value{
xxx
}
\description{
xxx
}

