% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ind.R
\name{create.ind}
\alias{create.ind}
\title{Create Indicator Matrix}
\usage{
create.ind(grp)
}
\arguments{
\item{grp}{A grouping vector.}
}
\description{
Create an indicator matrix.
}
\keyword{utility}
