% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.rs.R
\name{create.rs}
\alias{create.rs}
\title{Create a response style}
\usage{
create.rs(
  alpha = matrix(c(1, 2, 1), nrow = 1),
  nr.scale = 7,
  tvec = c(0, 0.5, 1),
  xvec = 0:nr.scale/nr.scale,
  scale = TRUE
)
}
\arguments{
\item{alpha}{vector of spline coefficients}

\item{nr.scale}{number of rating categories; numeric}

\item{tvec}{knots for spline functions}

\item{xvec}{evaluation points for basis functions}

\item{scale}{logical; scale or not}
}
\description{
Creates a response style by cutting up a quadratic monotone spline.
}
\author{
Pieter C. Schoonees
}
\keyword{splines}
