% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createcdsdata.R
\name{createcdsdata}
\alias{createcdsdata}
\title{Create a cdsdata Object}
\usage{
createcdsdata(x, q = NULL)
}
\arguments{
\item{x}{A data frame or matrix containing the data.}

\item{q}{Optional; the maximum rating category, so that the rating scale used for all 
items are \code{1:q}.}
}
\description{
Create a cdsdata object from a data frame or matrix.
}
\keyword{multivariate}
