\name{defaultProb}
\alias{defaultProb}
\title{Approximate the default probability at time t given the parSpread}
\usage{
defaultProb(parSpread, t, recoveryRate = 0.4)
}
\arguments{
  \item{parSpread}{in bps}

  \item{t}{in years}

  \item{recoveryRate}{in decimal. Default is 0.4.}
}
\description{
Approximate the default probability at time t given the
parSpread
}

