% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmat.R
\name{indmat}
\alias{indmat}
\title{Create an Indicator Matrix}
\usage{
indmat(grp.vec, K = length(unique(grp.vec)))
}
\arguments{
\item{grp.vec}{Numeric vector giving the group membership.}

\item{K}{Scalar indicating the number of groups. Defaults to the number of
unique elements in \code{grp.vec}.}
}
\description{
Creates an indicator matrix from a grouping vector.
}
