% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispline.R
\name{ispline}
\alias{ispline}
\title{Quadratic monotone spline basis function for given knots.}
\usage{
ispline(xvec, tvec = c(0, 0.5, 1), intercept = TRUE)
}
\arguments{
\item{xvec}{Vector at which to evaluate the basis functions.}

\item{tvec}{Vector of spline knots: lower endpoint, interior knot, upper endpoint.}

\item{intercept}{Logical; should an intercept be included or not?}
}
\description{
Calculate basis functions for monotone quadratic splines.
}
\keyword{multivariate}
