% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lfun.G.upd.R
\name{Lfun.G.upd}
\alias{Lfun.G.upd}
\title{Calculate Loss for G Update}
\usage{
Lfun.G.upd(G, a.cur, bwts2, Fr.bk, n, m, q, K)
}
\arguments{
\item{G}{The current value for G.}

\item{a.cur}{The current value for a.}

\item{bwts2}{The current value of the squared b weights.}

\item{Fr.bk}{Current product between Fr.cent and bk.}

\item{n}{Number of respondents.}

\item{m}{Number of items.}

\item{q}{Number for rating scale categories so that the rating scale is \code{1:q}.}

\item{K}{Number of response style groups.}
}
\description{
Loss function used for updating G. This is not equivalent to the original loss function, as 
only a part of the total loss depends on G.
}
\keyword{multivariate}
