\name{mondf}
\alias{mondf}
\title{compute a month difference as a difference between two monnb's}
\usage{
mondf(d1, d2)
}
\arguments{
  \item{d1}{date 1}

  \item{d2}{date 2}
}
\value{
month difference as a difference between two monnb's
}
\description{
compute a month difference as a difference between two
monnb's
}

