% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cdslist.R
\name{plot.cdslist}
\alias{plot.cdslist}
\title{Plot a \code{cdslist} Object}
\usage{
\method{plot}{cdslist}(x, which = 2L, ...)
}
\arguments{
\item{x}{An object of class \code{cdslist}.}

\item{which}{The which argument passed to \code{\link{plot.cds}}.}

\item{\dots}{Additional arguments passed to \code{\link{plot.cds}}.}
}
\description{
Create a scree plot and bubble plots for all elements in a \code{cdslist} object.
}
