% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cds.R
\name{print.cds}
\alias{print.cds}
\title{Print cds Object}
\usage{
\method{print}{cds}(x, ...)
}
\arguments{
\item{x}{A \code{cds} object.}

\item{\dots}{Unimplemented.}
}
\description{
Print method for \code{cds} objects.
}
\keyword{utility}
