% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cdsdata.R
\name{print.cdsdata}
\alias{print.cdsdata}
\title{Print dsdata Objects}
\usage{
\method{print}{cdsdata}(x, ...)
}
\arguments{
\item{x}{A \code{cdsdata} object}

\item{\dots}{Unimplemented.}
}
\description{
This is a simple print method  for object that inherits from the class \code{cdsdata}.
}
