\docType{methods}
\name{recRisk01-method}
\alias{recRisk01,CDS-method}
\alias{recRisk01-method}
\title{Calculate the amount of change in upfront when there is a 1\%
increase in recovery rate.}
\usage{
\S4method{recRisk01}{CDS}(object)
}
\arguments{
  \item{object}{the input \code{CDS} class object}
}
\description{
Calculate the amount of change in upfront when there is a
1\% increase in recovery rate.
}

