% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensory.aux.R
\docType{data}
\name{sensory.aux}
\alias{sensory.aux}
\title{Auxiliary Information for \code{\link{sensory}} Data}
\format{
A data frame with 268 observations on the following 3 variables.
\describe{ \item{Gender}{a factor with levels \code{F} for females
and \code{M} for males} \item{Age}{a factor for age with levels
\code{14 to 24},\code{25 to 34}, \code{35 to 44}, and \code{45 to 55}}
\item{Consumption}{a factor for consumption with levels \code{Heavy
consumer}, \code{Light consumer}, and \code{Medium consumer}} }
}
\source{
%% ~~ reference to a publication or URL from which the data were
obtained ~~
}
\description{
Auxiliary Information for \code{\link{sensory}} Data
}
\examples{
data(sensory.aux)

}
\keyword{datasets}
