% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensory.R
\docType{data}
\name{sensory}
\alias{sensory}
\title{sensory Data}
\description{
Data from 268 panellists rating each of 20 different products on 7 attributes. It is presented
in a \code{data.frame} with 268 observations on 140 variables. Each observation 
represents a different trained panellist. The columns correspond to products and items.
The 20 different products are coded by alphabetic letters from A to T, and the items are coded
from 1 to 7. So item C5 corresponds to product C being rated on item 5.
}
\examples{
data(sensory)

}
\keyword{datasets}
