\docType{methods}
\name{spreadDV01-method}
\alias{spreadDV01,CDS-method}
\alias{spreadDV01-method}
\title{The spreadDV01 method for CDS class.}
\usage{
\S4method{spreadDV01}{CDS}(object)
}
\arguments{
  \item{object}{the input \code{CDS} class object}
}
\description{
Calculate the \code{spreadDV01} from conventional spread
}

