\docType{methods}
\name{summary,CDS-method}
\alias{summary,CDS-method}
\title{The summary method displays only the essential info about the CDS
class object.}
\usage{
\S4method{summary}{CDS}(object, ...)
}
\arguments{
  \item{object}{an object for which a summary is desired.}

  \item{...}{additional arguments affecting the summary
  produced.}
}
\description{
The summary method displays only the essential info about
the CDS class object.
}

