% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trQnorm.R
\name{trQnorm}
\alias{trQnorm}
\title{Truncated Normal Quantiles}
\usage{
trQnorm(p, mean = 1, sd = 1, a = 0, b = 1)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{mean}{The mean of the distribution.}

\item{sd}{The standard deviation.}

\item{a}{Lower truncation point.}

\item{b}{Upper truncation point.}
}
\description{
Quantile function of the truncated normal distribution.
}
\author{
Pieter C. Schoonees
}
\keyword{multivariate}
