% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trRnorm.R
\name{trRnorm}
\alias{trRnorm}
\title{Truncated Normal Sampling}
\usage{
trRnorm(n, mu = 0, sd = 1, a = -Inf, b = Inf)
}
\arguments{
\item{n}{The number of points to sample.}

\item{mu}{The mean of the distribution.}

\item{sd}{The standard deviation.}

\item{a}{The lower truncation point.}

\item{b}{The upper truncation point.}
}
\description{
Random numbers from truncated univariate normal.
}
\keyword{multivariate}
