% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS.R
\name{CDS}
\alias{CDS}
\title{Check for geographic centroids}
\usage{
CDS(coordinates, maxdist = NULL, maxdistrel = NULL)
}
\arguments{
\item{coordinates}{Data.frame or matrix containing two columns: latitude and longitude (in that order).}

\item{maxdist}{Maximum distance from centroid to qualify as centroid}

\item{maxdistrel}{Maximum relative distance from centroid (relative to distance from centroid  to farthest point in political division), to qualify as a centroid}
}
\value{
Data.frame containing CDS output.
}
\description{
Submits latitude/longitude coordinates to the centroid detection service
}
\examples{
{
# Test file from BIEN website:
coordinates <-
read.csv("https://bien.nceas.ucsb.edu/bien/wp-content/uploads/2020/10/cds_testfile.csv",
stringsAsFactors = FALSE)
CDS_output <- CDS(coordinates)
}

}
